Move Inputs to ConfigView folder
