+++ FtcDashboard/dash/src/containers/GraphView/GraphView.tsx
@@ -11,7 +11,7 @@ import BaseView, {
 } from '../BaseView';
 import MultipleCheckbox from '../../components/MultipleCheckbox';
 import GraphCanvas from './GraphCanvas';
-import TextInput from '../../components/inputs/TextInput';
+import TextInput from '../ConfigView/inputs/TextInput';
 
 import { ReactComponent as ChartIcon } from '../../assets/icons/chart.svg';
 import { ReactComponent as CloseIcon } from '../../assets/icons/close.svg';
@@ -19,7 +19,7 @@ import { ReactComponent as PlayIcon } from '../../assets/icons/play_arrow.svg';
 import { ReactComponent as PauseIcon } from '../../assets/icons/pause.svg';
 
 import { RootState } from '../../store/reducers';
-import { validateInt } from '../../components/inputs/validation';
+import { validateInt } from '../ConfigView/inputs/validation';
 import { DEFAULT_OPTIONS } from './Graph';
 
 type GraphViewState = {
